+++ NEWS.md
@@ -126,6 +126,10 @@ Deprecated or removed
 
   * Deprecate `chol(A,Val{:U/:L})` in favor of `chol(A)` ([#13680]).
 
+  * `issym` is deprecated in favor of `issymmetric` to match similar functions (`ishermitian`, ...) ([#15192])
+
+  * `scale` is deprecated in favor of either `α*A`, `Diagonal(x)*A`, or `A*Diagonal(x)`. ([#15258])
+
 Julia v0.4.0 Release Notes
 ==========================